+++ dns/config.go
@@ -47,6 +47,9 @@ func (c *Config) Client() (interface{}, error) {
 		return nil, fmt.Errorf("Error configuring provider: when using authentication, \"key_name\", \"key_secret\" and \"key_algorithm\" should be non empty")
 	}
 	client.c = new(dns.Client)
+	client.transport = c.transport
+	client.c.Timeout = c.timeout
+	client.retries = c.retries
 	if c.keyname != "" {
 		if !dns.IsFqdn(c.keyname) {
 			return nil, fmt.Errorf("Error configuring provider: \"key_name\" should be fully-qualified")